requirements.txt
streamlit==1.27.0
tensorflow==2.15.0
numpy
pandas
rdkit-pypi==2023.9.5
sascorer